refactor: timer가 필요한 부분 컴포넌트로 분리하여 리렌더링 영역 최소화
